Remane all package to remove "package" subfix
