+++ src/main/java/tutorlink/appstate/AppState.java
@@ -1,0 +1,9 @@
+package tutorlink.appstate;
+
+import tutorlink.lists.AssignmentList;
+import tutorlink.lists.StudentList;
+
+public class AppState {
+    public StudentList students = new StudentList();
+    public AssignmentList assignments = new AssignmentList();
+}
