--- src/main/java/tutorlink/appstatepackage/AppState.java
@@ -1,9 +1,0 @@
-package tutorlink.appstatepackage;
-
-import tutorlink.listpackage.AssignmentList;
-import tutorlink.listpackage.StudentList;
-
-public class AppState {
-    public StudentList students = new StudentList();
-    public AssignmentList assignments = new AssignmentList();
-}
