+++ src/main/java/tutorlink/component/ClassParticipation.java
@@ -1,0 +1,11 @@
+package tutorlink.component;
+
+public class ClassParticipation extends Component{
+    public ClassParticipation (String name, double maxScore, double weight) {
+        super(name, maxScore, weight);
+    }
+    @Override
+    public String toString() {
+        return "ClassParticipation " + super.toString();
+    }
+}
