+++ src/main/java/tutorlink/component/Exam.java
@@ -1,0 +1,11 @@
+package tutorlink.component;
+
+public class Exam extends Component{
+    public Exam(String name, double maxScore, double weight) {
+        super(name, maxScore, weight);
+    }
+    @Override
+    public String toString() {
+        return "Exam " + super.toString();
+    }
+}
