--- src/main/java/tutorlink/componentpackage/Assignment.java
@@ -1,12 +1,0 @@
-package tutorlink.componentpackage;
-
-public class Assignment extends Component {
-    public Assignment(String name, double maxScore, double weight) {
-        super(name, maxScore, weight);
-    }
-
-    @Override
-    public String toString() {
-        return "Assignment " + super.toString();
-    }
-}
