--- src/main/java/tutorlink/componentpackage/ClassParticipation.java
@@ -1,11 +1,0 @@
-package tutorlink.componentpackage;
-
-public class ClassParticipation extends Component{
-    public ClassParticipation (String name, double maxScore, double weight) {
-        super(name, maxScore, weight);
-    }
-    @Override
-    public String toString() {
-        return "ClassParticipation " + super.toString();
-    }
-}
