--- src/main/java/tutorlink/componentpackage/Exam.java
@@ -1,11 +1,0 @@
-package tutorlink.componentpackage;
-
-public class Exam extends Component{
-    public Exam(String name, double maxScore, double weight) {
-        super(name, maxScore, weight);
-    }
-    @Override
-    public String toString() {
-        return "Exam " + super.toString();
-    }
-}
