+++ src/main/java/tutorlink/exceptions/DuplicateMatricNumberException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public class DuplicateMatricNumberException extends TutorLinkException {
+    public DuplicateMatricNumberException(String message) {
+        super(message);
+    }
+}
