+++ src/main/java/tutorlink/exceptions/IllegalValueException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public class IllegalValueException extends TutorLinkException {
+    public IllegalValueException(String message) {
+        super(message);
+    }
+}
