+++ src/main/java/tutorlink/exceptions/InvalidWeightingException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public class InvalidWeightingException extends TutorLinkException {
+    public InvalidWeightingException(String message) {
+        super(message);
+    }
+}
