+++ src/main/java/tutorlink/exceptions/ItemNotFoundException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public class ItemNotFoundException extends TutorLinkException {
+    public ItemNotFoundException(String message) {
+        super(message);
+    }
+}
