+++ src/main/java/tutorlink/exceptions/MissingNextLineException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public class MissingNextLineException extends TutorLinkException {
+    public MissingNextLineException(String message) {
+        super(message);
+    }
+}
