+++ src/main/java/tutorlink/exceptions/StudentNotFoundException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public class StudentNotFoundException extends TutorLinkException {
+    public StudentNotFoundException(String message) {
+        super(message);
+    }
+}
