+++ src/main/java/tutorlink/exceptions/TutorLinkException.java
@@ -1,0 +1,7 @@
+package tutorlink.exceptions;
+
+public abstract class TutorLinkException extends RuntimeException {
+    public TutorLinkException(String message) {
+        super(message);
+    }
+}
