--- src/main/java/tutorlink/exceptionspackage/DuplicateMatricNumberException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public class DuplicateMatricNumberException extends TutorLinkException {
-    public DuplicateMatricNumberException(String message) {
-        super(message);
-    }
-}
