--- src/main/java/tutorlink/exceptionspackage/IllegalValueException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public class IllegalValueException extends TutorLinkException {
-    public IllegalValueException(String message) {
-        super(message);
-    }
-}
