--- src/main/java/tutorlink/exceptionspackage/InvalidWeightingException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public class InvalidWeightingException extends TutorLinkException {
-    public InvalidWeightingException(String message) {
-        super(message);
-    }
-}
