--- src/main/java/tutorlink/exceptionspackage/ItemNotFoundException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public class ItemNotFoundException extends TutorLinkException {
-    public ItemNotFoundException(String message) {
-        super(message);
-    }
-}
