--- src/main/java/tutorlink/exceptionspackage/MissingNextLineException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public class MissingNextLineException extends TutorLinkException {
-    public MissingNextLineException(String message) {
-        super(message);
-    }
-}
