--- src/main/java/tutorlink/exceptionspackage/StudentNotFoundException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public class StudentNotFoundException extends TutorLinkException {
-    public StudentNotFoundException(String message) {
-        super(message);
-    }
-}
