--- src/main/java/tutorlink/exceptionspackage/TutorLinkException.java
@@ -1,7 +1,0 @@
-package tutorlink.exceptionspackage;
-
-public abstract class TutorLinkException extends RuntimeException {
-    public TutorLinkException(String message) {
-        super(message);
-    }
-}
