--- src/main/java/tutorlink/listpackage/ItemList.java
@@ -1,5 +1,0 @@
-package tutorlink.listpackage;
-
-public class ItemList {
-
-}
