+++ src/main/java/tutorlink/result/CommandResult.java
@@ -1,0 +1,17 @@
+package tutorlink.result;
+
+public class CommandResult {
+
+    private static final String LINE_SEPARATOR = "\n\t";
+
+    public final String message;
+
+    public CommandResult(String message) {
+        this.message = message;
+    }
+
+    @Override
+    public String toString() {
+        return message;
+    }
+}
