--- src/main/java/tutorlink/resultpackage/CommandResult.java
@@ -1,17 +1,0 @@
-package tutorlink.resultpackage;
-
-public class CommandResult {
-
-    private static final String LINE_SEPARATOR = "\n\t";
-
-    public final String message;
-
-    public CommandResult(String message) {
-        this.message = message;
-    }
-
-    @Override
-    public String toString() {
-        return message;
-    }
-}
